requirements.txt
adrf==0.1.2
asgiref==3.7.2
async-property==0.2.2
attrs==23.1.0
click==8.1.7
colorama==0.4.6
Django==4.2.7
django-environ==0.11.2
django-filter==23.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
factory-boy==3.3.0
Faker==20.0.3
h11==0.14.0
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
packaging==23.2
pluggy==1.3.0
psycopg2-binary
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.13.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
uvicorn==0.24.0.post1
